moved completed books into its own component
